***NOTE*** Do not alter any of the file paths.
The required files are in the required folders. Tested.
All pages work including registration page in php.

This is my part for the project.

I have updated the webiste using bootstrap 4 templates. Added admin.php a file that the user is redirected to after logging in and shows an admin dashboard.